Misc - Fixed MSVC compiler warnings

+++ source/core/logger.cpp
@@ -29,13 +29,13 @@ void GlobalLogger::Write( const char* pFormat, ... )
     va_list argList;
     va_start( argList, pFormat );
 
-    size_t bufferOutputSize = vsnprintf( pBuffer, sizeof( pBuffer ), pFormat, argList );
+    int32_t bufferOutputSize = vsnprintf( pBuffer, sizeof( pBuffer ), pFormat, argList );
     OTDU_ASSERT( bufferOutputSize < sizeof( pBuffer ) );
 
     va_end( argList );
 
     // Bad format/buffer overrun; early exit to avoid hard crash.
-    if ( bufferOutputSize == -1ull ) {
+    if ( bufferOutputSize == -1 ) {
         OTDU_TRIGGER_BREAKPOINT;
         return;
     }
